src/requirements.txt
Flask==3.1.0
pymongo==4.10.1
python-dotenv==1.0.1
Werkzeug==3.1.3
yt_dlp==2025.1.15
gunicorn==23.0.0
flask_httpauth==4.8.0